Project working after merging.

+++ src/main/java/org/ucsc/enmoskill/Services/SupportDELETE.java
@@ -1,6 +1,7 @@
 package org.ucsc.enmoskill.Services;
 
 import org.ucsc.enmoskill.database.DatabaseConnection;
+import org.ucsc.enmoskill.utils.TokenService;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
@@ -9,21 +10,26 @@ import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
 public class SupportDELETE {
-    private String requesterID;
-    public SupportDELETE(String requesterID, HttpServletResponse response) throws SQLException, IOException {
+    private String TicketID;
+    private HttpServletResponse response;
+    private TokenService.TokenInfo tokenInfo;
+    public SupportDELETE(String TicketID,TokenService.TokenInfo tokenInfo, HttpServletResponse response) throws SQLException, IOException {
         Connection connection = DatabaseConnection.initializeDatabase();
-        String quary = "DELETE FROM enmo_database.ticket WHERE ref_no = " + requesterID;
+
+        String quary = "DELETE FROM enmo_database.ticket WHERE ref_no = " + TicketID+" AND requesterID = "+tokenInfo.getUserId();
+
         PreparedStatement preparedStatement = connection.prepareStatement(quary);
 
         int rowsAffected = preparedStatement.executeUpdate();
 
         if (rowsAffected > 0) {
 
-            response.getWriter().write("Row is deleted successfully.");
+            response.getWriter().write("Ticket deleted successfully");
             response.setStatus(HttpServletResponse.SC_OK);
         } else {
 
-            response.getWriter().write("No row found .");
+            response.getWriter().write("Not found ticket.");
+            response.setStatus(HttpServletResponse.SC_NOT_FOUND);
         }
     }
 }
